docker-images/fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
psycopg2-binary==2.9.9
python-multipart==0.0.6